requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
amqp==5.2.0
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo[tzdata]==0.2.1
    # via
    #   -r requirements/base.txt
    #   backports-zoneinfo
    #   celery
    #   kombu
billiard==4.2.0
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.3.6
    # via -r requirements/base.txt
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
coverage[toml]==7.3.2
    # via
    #   coverage
    #   pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements/base.txt
edx-opaque-keys==2.5.1
    # via -r requirements/base.txt
exceptiongroup==1.2.0
    # via pytest
iniconfig==2.0.0
    # via pytest
kombu==5.3.4
    # via
    #   -r requirements/base.txt
    #   celery
openedx-atlas==0.5.0
    # via -r requirements/base.txt
packaging==23.2
    # via pytest
pbr==6.0.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.41
    # via
    #   -r requirements/base.txt
    #   click-repl
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pytest==7.4.3
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.7.0
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   celery
pytz==2023.3.post1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.8.0
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   edx-opaque-keys
    #   kombu
tzdata==2023.3
    # via
    #   -r requirements/base.txt
    #   backports-zoneinfo
    #   celery
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.12
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
